feat: implement shell input/output bindings and tree structure visualization
