+++ shell/tree_generator.go
@@ -4,9 +4,11 @@ import (
 	"fmt"
 	"reflect"
 
+	"github.com/fmotalleb/scrapper-go/utils"
 	"github.com/rivo/tview"
 )
 
+// newNode creates a new tree node with the given label and collapsible behavior.
 func newNode(label string, collapsible bool) *tview.TreeNode {
 	node := tview.NewTreeNode(label).
 		SetTextStyle(outputTheme).
@@ -21,43 +23,38 @@ func newNode(label string, collapsible bool) *tview.TreeNode {
 	return node
 }
 
+// buildTree constructs a tree from the provided data and returns the root node.
 func buildTree(data any) *tview.TreeNode {
 	root := newNode("root", false)
 	buildSubTree(root, data)
 	return root
 }
 
-// buildSubTree of the parent node with data
-func buildSubTree(node *tview.TreeNode, data any) {
+// buildSubTree recursively builds the tree for a given parent node and data.
+func buildSubTree(node *tview.TreeNode, data any, replaceLabel ...bool) {
 	val := reflect.ValueOf(data)
+
 	switch val.Kind() {
 	case reflect.Map:
-		// Iterate over map keys
+		// Iterate over map keys and build subtree for each key-value pair.
 		for _, key := range val.MapKeys() {
-			// Get the value corresponding to the key
 			val := val.MapIndex(key).Interface()
-			// Create a new node for the key
 			subNode := newNode(fmt.Sprintf("%v: %v", key.Interface(), val), true)
-			// Recursively build the subtree for the value
-			if _, ok := val.(string); !ok {
-				buildSubTree(subNode, val)
-			}
-
-			// Add the subNode to the parent node
+			// If the value is not a string, build a sub-tree for it.
+			buildSubTree(subNode, val, false)
 			node.AddChild(subNode)
 		}
 	case reflect.Slice, reflect.Array:
-		// Iterate over each element in the slice/array
+		// Iterate over each element in the slice/array and build a subtree for each.
 		for i := 0; i < val.Len(); i++ {
-			// Create a new node for each index
 			subNode := newNode(fmt.Sprintf("[%d]", i), true)
-			// Recursively build the subtree for each element
 			buildSubTree(subNode, val.Index(i).Interface())
-			// Add the subNode to the parent node
 			node.AddChild(subNode)
 		}
 	default:
-		// For non-map, non-slice types, set the text as the value itself
-		node.SetText(fmt.Sprintf("%v", data))
+		// For primitive types, set the text as the value itself.
+		if utils.FirstOr(replaceLabel, true) {
+			node.SetText(fmt.Sprintf("%v", data))
+		}
 	}
 }
